requirements.txt
imageio==2.35.1
numpy==2.1.0
pillow==10.4.0
